ANN2SNN/requirements.txt
absl_py==1.3.0
imageio==2.9.0
matplotlib==3.5.1
numpy==1.21.5
scipy==1.7.3
tensorboardX==2.2
tensorboardX==2.6.2.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
